pyproject.toml
[tool.poetry]
name = "pocket-assistant"
version = "0.1.0"
description = "Telegram bot with some useful features"
authors = ["quantumwaffy <sashchystyh@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.5"
alembic = "^1.12.0"
SQLAlchemy = "^2.0.20"
asyncpg = "^0.28.0"
aiogram = "^3.0.0"
pydantic-settings = "^2.0.3"
websockets = "^11.0.3"
pymongo = "^4.5.0"
motor = "^3.3.1"
beanie = "^1.22.6"
broadcaster = {extras = ["redis"], version = "^0.2.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pytest = "^7.4.1"
httpx = "^0.24.1"
aiosqlite = "^0.19.0"
pytest-mock = "^3.11.1"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autoflake]
recursive = true

[tool.black]
line-length = 120
target-version = ["py311"]
exclude = "^migrations/"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

